+++ arkham/src/main/java/co/com/cesde/arkham/persistence/crud/AdministradorJpaRepository.java
@@ -4,7 +4,8 @@ import co.com.cesde.arkham.persistence.entity.Administrador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface AdministradorJpaRepository extends JpaRepository<Administrador, Long> {
+public interface AdministradorJpaRepository extends JpaRepository<Administrador, Integer> {
     List<Administrador> findByNombreAdministrador(String nombre);
 }
