+++ arkham/src/main/java/co/com/cesde/arkham/persistence/crud/InmuebleJpaRepository.java
@@ -1,15 +1,13 @@
 package co.com.cesde.arkham.persistence.crud;
 
-import co.com.cesde.arkham.persistence.entity.Barrio;
 import co.com.cesde.arkham.persistence.entity.Inmueble;
 import co.com.cesde.arkham.persistence.entity.Propietario;
 import co.com.cesde.arkham.persistence.entity.Ubicacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
-public interface InmuebleJpaRepository extends JpaRepository<Inmueble,Long> {
+public interface InmuebleJpaRepository extends JpaRepository<Inmueble,Integer> {
     Inmueble findByUbicacion(Ubicacion ubicacion);
 
     List<Inmueble> findByPropietario(Propietario propietario);
