+++ arkham/src/main/java/co/com/cesde/arkham/persistence/crud/PropietarioJpaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface PropietarioJpaRepository extends JpaRepository<Propietario,Long> {
+public interface PropietarioJpaRepository extends JpaRepository<Propietario,Integer> {
 
     List<Propietario> findByNombrePropietario(String nombrePropietario);
 }
