+++ arkham/src/main/java/co/com/cesde/arkham/persistence/crud/UsuarioJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuarioJpaRepository extends JpaRepository<Usuario,Long> {
+public interface UsuarioJpaRepository extends JpaRepository<Usuario,Integer> {
     Optional<Usuario> findByUsuario(String usuario);
 }
